Refactors release workflow for multiple packages
Migrates to a single release workflow file to handle the release of multiple packages (nextmv, nextmv-gurobipy, and nextmv-sklearn).

This change eliminates the need for a separate publish-notify workflow. Instead, the publishing and notification steps are now directly integrated into the main release workflow, simplifying the release process and making it more maintainable.